Tab bar works, plus image holder and take picture btn
Tab bar changes class on the tabItems so the tab clicked will be the new active tab.  Right now it just puts #TakePic, #uploadPic or #help in the url, this is going to be changed to the corrosponding html files for the views.
The image holder has been made with the take picture button. So now we just need the function in those.